src/main/java/yome/fgo/simulator/models/effects/GrantBuff.java
package yome.fgo.simulator.models.effects;

import lombok.experimental.SuperBuilder;
import yome.fgo.data.proto.FgoStorageData.BuffData;
import yome.fgo.data.proto.FgoStorageData.Target;
import yome.fgo.simulator.models.Simulation;
import yome.fgo.simulator.models.combatants.Combatant;
import yome.fgo.simulator.models.effects.buffs.BuffFactory;
import yome.fgo.simulator.utils.TargetUtils;

import java.util.List;

@SuperBuilder
public class GrantBuff extends Effect {
    private List<BuffData> buffData;
    private Target target;
    private int buffLevel;

    @Override
    public void apply(Simulation simulation, int level) {
        for (final Combatant combatant : TargetUtils.getTargets(simulation, target)) {
            simulation.setEffectTarget(combatant);
            if (shouldApply(simulation)) {
                combatant.addBuff(BuffFactory.buildBuff(buffData.get(level - 1), buffLevel));
            }
            simulation.setEffectTarget(null);
        }
    }
}
